Rebuilt how the regeneration region works (#4)
RegenRegions can now be splittet into several smaller regions. This allows to be more flexible: Big Regions with eg. 4 million blocks are crashing the server. If only small fitting regions are set, there is no need for 4 million blocks to regen.
Also a block limit is implemented to ensure that a crash is not going to happen. This limit can be changed within the minigames config file and is default set to 300k.